Clean up loaders by not passing None to `options`.
This saves implementation of every `Loader.load()` from having

```
if options is None:
    options = {}
```
