+++ pydax/loaders/_base.py
@@ -16,7 +16,7 @@
 
 from abc import ABC, abstractmethod
 
-from typing import Any, Dict, Optional, Union
+from typing import Any, Dict, Union
 
 from .. import _typing
 from .._schema import SchemaDict
@@ -26,8 +26,8 @@ class Loader(ABC):
     "Base class of all loaders."
 
     @abstractmethod
-    def load(self, path: Union[_typing.PathLike, Dict[str, str]], options: Optional[SchemaDict]) -> Any:
-        """Loads from a give path or a dict of path configurations. This should must overridden when inherited.
+    def load(self, path: Union[_typing.PathLike, Dict[str, str]], options: SchemaDict) -> Any:
+        """Loads from a give path or a dict of path configurations. This must be overridden when inherited.
 
         :param path: The path or path configurations of the files to be loaded.
         :param options: Options passed to the loader.
